refactor storage structure
