+++ minigu/storage/src/common/mod.rs
@@ -1,0 +1,10 @@
+pub mod iterators;
+pub mod model;
+pub mod wal;
+
+// Re-export commonly used types
+pub use model::edge::*;
+pub use model::properties::*;
+pub use model::schema::*;
+pub use model::vertex::*;
+pub use wal::*;
