+++ minigu/storage/src/common/model/edge.rs
