+++ minigu/storage/src/common/model/properties.rs
