+++ minigu/storage/src/common/model/vertex.rs
