--- minigu/storage/src/memory/mod.rs
@@ -1,6 +1,0 @@
-pub mod adjacency_iterator;
-pub mod checkpoint;
-pub mod edge_iterator;
-pub mod memory_graph;
-pub mod transaction;
-pub mod vertex_iterator;
